gtk: Fixed multi-monitor patch to apply only to non-top-level windows. Top-level windows are already positioned arbitrarily on the primary monitor and don't need the correction.
Reported in https://github.com/hexchat/hexchat/issues/1007 by @Bill48105 and @emko. The former's bug was about the positioning of the Network List window on first-run with an empty profile. The latter's bug was for the file open dialog popped up via gtk_fileutil_req. Both of these are top-level windows and were thus affected.