Switch to upstream harfbuzz + Fan's MSVC makefiles
Use our own fork of harfbuzz which is a released tarball of the
upstream version plus Fan's MSVC patches. Hopefully these will be
merged soon and we can drop the fork.
We still patch the makefile to remove the -vs$(VSVER) suffix from
the dll names.

This also removes ragel from the dependencies: upstream tarballs
already contain the ragel generated files in the distribution.