protobuf: build Release instead of RelWithDebugInfo
The reason for this is that the protobuf build does not properly
support generating the pdb files and we endup with vc120.pdb files
instead of the ones with the proper name. Let's not use it to avoid
problems with the build later.

If in the future protobuf adds this feature we can add back the pdb
file generation.