Add option --from-scratch to the build to remove the temporary dirs.
It cleans completely the build, the destination and the tool dir.
The difference from --clean is that also the destination dir (for example
c:\gtk-build\gtk\win32\release) is removed so there are less mistake (missing
dependency already present on disk).
To keep the tools, common for all the platform/configuration, you can pass the
--keep-tool option.