Update libxml2 and openssl to use http instead of ftp
Behind corporate firewalls, it is very difficult to access ftp locations. I couldn't get it to access the ftp based projects even with ftp_proxy set as an environmental variables and in Internet Options in Windows. This PR updates the two projects that were using the ftp protocol to use http and https instead.