Handle the Vs2017 version (professional, community, ...) automatically
For the Visual Studio 2017, look for Professional, BuildTools, Enterprise or Community subdirectory under 'c:\Programs Files (x86)\Microsoft Visual Studio\2017' if the directory is not set on the command line and enable the first found.

If there is some errors and the visual studio is not found try to execute the vswhere.exe program (that comes with the new Visual Studio installer) to
locate the installations path and display them to the user.

Resolve #252