Automatically load the win-sdk-ver value.
If not set on the command line the windows sdk version is calculated (from the
vs environment) and loaded so it can be used in the msbuild command.
This avois errors on the vs solutions where there is a sdk version that is not installed (you
should install all the sdk to avoid the problem or fix all the solutions).