Added ms_build_gen(...) to llok for the newest .sln file
From a solutions base directory (e.g. build\win32) and the solution
file (e.g. gtk+.sln) the scripts looks for vsXX or vsYYYY (15 or
2017 for vs2017) to locate the .sln file.
If it's not found go back with the version/year (vs14/vs2015, then
vs12/vs2013) until it founds one or fail.
So if you have only the vs2015 solution you can build the project
automatically with the vs2017.
This is been done to have the test the vs2017 support.