Download with nuget the python used to build the projects.
The python that launch the script is used only for the script and can be
any 3.6+ version, even the embedded one.
The python needed by the projects is downloaded and installed with nuget,
respecting the win32/x64 architecture.
You can use any of 3.5, 3.6, 3.7 or a beta version of the new 3.8.