Fix Visual Studio Version Selection
Fix how the version of visual studio is selected.

Visual Studio is selected by checking the path for the presence of
either the year or the version number. Unfortunately, some version
number are present in different editions. This can lead to the wrong
installed version being selected. For example, Visual Studio 2022 has a
version of 17, so specifying the use of version 17 can result in 2017
being selected instead.