ffmpeg: Enable avfilter and avformat for gst-libav
Previously, gst-libav failed to build due to absence of avfilter and
avformat libraries from ffmpeg. This commit enables those libraries by
fixing the build error with unistd.h. This imports a patch for zlib
from vcpkg[^1] to fix the unistd.h check.

In gvsbuild, zlib is build using nmake which copies the zconf.h file.
So, the imported patch file was modified for zconf.h only.

[^1]: https://github.com/microsoft/vcpkg/blob/master/ports/zlib/0001-Prevent-invalid-inclusions-when-HAVE_-is-set-to-0.patch