Improve error message when Visual Studio can't be found
This addresses the confusion of `gvsbuild` showing that it _found_ Visual Studio installations, but then logging right after that it couldn't, in fact, find it.

This adds logs for each VS Installation that it considers "inapplicable", such as because:
* The version doesn't match `vs_ver`
* The `vcvars.bat` is missing

-----

The `exit_missing` parameter was removed because it was _always_ set to `False`.